Move SrcLocation to parser
