+++ src/AstToRamTranslator.cpp
@@ -21,7 +21,6 @@
 #include "Global.h"
 #include "LogStatement.h"
 #include "RelationTag.h"
-#include "SrcLocation.h"
 #include "ast/Aggregator.h"
 #include "ast/Argument.h"
 #include "ast/Atom.h"
@@ -62,6 +61,7 @@
 #include "ast/analysis/TopologicallySortedSCCGraph.h"
 #include "ast/analysis/TypeEnvironment.h"
 #include "ast/analysis/TypeSystem.h"
+#include "parser/SrcLocation.h"
 #include "ram/Condition.h"
 #include "ram/Expression.h"
 #include "ram/Node.h"
