+++ src/ast/Atom.h
@@ -21,7 +21,7 @@
 #include "Node.h"
 #include "NodeMapper.h"
 #include "QualifiedName.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "utility/ContainerUtil.h"
 #include "utility/StreamUtil.h"
 #include <algorithm>
