+++ src/ast/BooleanConstraint.h
@@ -18,7 +18,7 @@
 
 #include "Constraint.h"
 #include "Node.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include <cassert>
 #include <iostream>
 #include <string>
