+++ src/ast/Clause.h
@@ -21,7 +21,7 @@
 #include "Literal.h"
 #include "Node.h"
 #include "NodeMapper.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "utility/ContainerUtil.h"
 #include "utility/MiscUtil.h"
 #include "utility/StreamUtil.h"
