+++ src/ast/ComponentInit.h
@@ -19,7 +19,7 @@
 #include "ComponentType.h"
 #include "Node.h"
 #include "NodeMapper.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "utility/MiscUtil.h"
 #include <memory>
 #include <ostream>
