+++ src/ast/IntrinsicFunctor.h
@@ -20,7 +20,7 @@
 #include "Functor.h"
 #include "FunctorOps.h"
 #include "Node.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "TypeAttribute.h"
 #include "utility/ContainerUtil.h"
 #include "utility/StreamUtil.h"
