+++ src/ast/RecordType.h
@@ -19,7 +19,7 @@
 #include "Attribute.h"
 #include "Node.h"
 #include "QualifiedName.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "Type.h"
 #include "utility/ContainerUtil.h"
 #include "utility/StreamUtil.h"
