+++ src/ast/SubroutineArgument.h
@@ -18,7 +18,7 @@
 
 #include "Argument.h"
 #include "Node.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include <cstddef>
 #include <ostream>
 #include <string>
