+++ src/ast/SubsetType.h
@@ -18,7 +18,7 @@
 
 #include "Node.h"
 #include "QualifiedName.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "Type.h"
 #include <iostream>
 #include <string>
