+++ src/ast/Term.h
@@ -19,7 +19,7 @@
 #include "Argument.h"
 #include "Node.h"
 #include "NodeMapper.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "utility/ContainerUtil.h"
 #include <algorithm>
 #include <memory>
