+++ src/ast/TypeCast.h
@@ -20,7 +20,7 @@
 #include "Node.h"
 #include "NodeMapper.h"
 #include "QualifiedName.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "utility/ContainerUtil.h"
 #include "utility/MiscUtil.h"
 #include <memory>
