+++ src/ast/UserDefinedFunctor.h
@@ -18,7 +18,7 @@
 
 #include "Functor.h"
 #include "Node.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include "TypeAttribute.h"
 #include "utility/ContainerUtil.h"
 #include "utility/MiscUtil.h"
