+++ src/ast/Variable.h
@@ -18,7 +18,7 @@
 
 #include "Argument.h"
 #include "Node.h"
-#include "SrcLocation.h"
+#include "parser/SrcLocation.h"
 #include <ostream>
 #include <string>
 #include <utility>
