+++ src/tests/ast_print_test.cpp
@@ -19,7 +19,6 @@
 #include "AggregateOp.h"
 #include "DebugReport.h"
 #include "ErrorReport.h"
-#include "parser/ParserDriver.h"
 #include "ast/Argument.h"
 #include "ast/Atom.h"
 #include "ast/Clause.h"
@@ -28,6 +27,7 @@
 #include "ast/Program.h"
 #include "ast/TranslationUnit.h"
 #include "ast/Variable.h"
+#include "parser/ParserDriver.h"
 #include <algorithm>
 #include <iosfwd>
 #include <memory>
