+++ src/tests/ast_transformers_test.cpp
@@ -18,7 +18,6 @@
 
 #include "DebugReport.h"
 #include "ErrorReport.h"
-#include "parser/ParserDriver.h"
 #include "ast/Clause.h"
 #include "ast/Node.h"
 #include "ast/Program.h"
@@ -27,6 +26,7 @@
 #include "ast/transform/MinimiseProgram.h"
 #include "ast/transform/RemoveRelationCopies.h"
 #include "ast/transform/ResolveAliases.h"
+#include "parser/ParserDriver.h"
 #include "utility/StringUtil.h"
 #include <memory>
 #include <string>
